requirements.txt
openai==1.42.0
playwright==1.46.0
WPP-Whatsapp==0.4.8.1
requests==2.32.3
pillow==10.4.0
opencv-python==4.10.0.84
click==8.1.7